Fixes for outputs of disabled pipelines.
The output of a statically-disabled pipeline is null.  However, the
output for a dynamically-disabled pipeline was not null - any constant
outputs would be present, as would any structure up to the point where a
reference to a disabled stage was enountered.  This is now fixed.

Various other fixes to interactions between disabled calls and map
calls, including better handling of mapped calls which are implicitly
disabled by mapping over an empty set.

Do not allow binding of the disabled control parameter to the output of
a stage which might be disabled, unless we can prove that the parent
pipeline of the stage in question will always be disabled if the target
stage is disabled.  Previously this was allowed and would result in the
stage not being disabled, which was perhaps surprising behavior.

Introduce a new runtime check that the value bound for the disabled
control is not null.  The previous behavior was to treat null as false,
which was surprising.  The new behavior is that null results in an
error, unless the (grand)parent pipeline was disabled.