Make a compile error for refs in untyped maps.
Previously, it was not possible to include a reference (e.g. self.foo)
in a map literal.  It is now, however, possible to include one in a
struct literal, and then to assign the struct to an untyped map.
Unfortunately this causes problems at runtime, because the type
information is required at that point and untyped maps don't have type
information.

This fixes things so rather than panicing at runtime, it will instead
fail when resolving the call graph, with an easier-to-understand error
message.