Impliment ErrorList.As() and Is().
Per the documentation for errors.As in go 1.13,

> An error matches target if the error's concrete value is assignable to the
> value pointed to by target, or if the error has a method As(interface{}) bool
> such that As(target) returns true.

This searches through the list of errors until one is found which
matches the target.  Is has similar behavior.

Unwrap is also implemented, returning the first non-nil error in the
list.

Also improve the documentation for ErrorList, and reduce memory
allocations in the common case of a singlton list.