Combine mrc, mrf, mrdr into single tool, mro.
At least for now, no changes to how they actually function, other than
that one can get to them as `mro <check|format|edit>`, respectively.
Each one is still entirely responsible for its own command-line parsing.
The new combined `mro` tool knows how to invoke the appropriate tool if
it is run from a symlink to one of the old names, for convenience.

The new combined tool's compiled size is 5479679 bytes, compared to
5014288, 4865692, and 5120100 bytes (respectively) for each of the old
binaries, or a total of 15000080 bytes, so this significantly reduces
the size of the distributed package (most of their size is the go
runtime and the syntax package, which they shared).