Don't populate file arrays with strings.
This was a bug in martian 2, and has been deprecated since martian 3.0.
The behavior was retained by default, however, for backwards
compatibility.  The new (correct) behavior, which was already the case
when running with `--strict=error` or for stages using any newer
features such as map calls or structs, is to pre-populate the output
with an empty array.