Avoid premature symlink resolution finding exe.
Particularly when running within a bazel runfiles tree, resolving
symlinks on the executable breaks out of the runfiles symlink forest,
with the result that other expected items might not be in the exepected
relative locations.  Attempt to use path.Dir(exec.LookPath(os.Args[0]))
instead, if either adapter or jobmanagers are in the expected relative
location from there.