Make paths MROPATH-relative in _invocation.
When generating _invocation files, use the absolute path when computing
the path to use for `@include`ing the source, because the original
include path might have been relative to another mro file that was not
in the MROPATH.

However, if the original input path was not in the MROPATH, then we
shouldn't convert relative paths to absolute paths, because the path was
probably intended to be relative to the working directory, and changing
the path in that case might mess things up if the MROPATH for, for
example, `mrg`, was not the same as the one for `mrp` later on.

Add more test cases for IncludeFilePath.

Add a docstring for Src.FindPath().