Fix some bugs with map-called pipelines.
If a stage depends for its inputs on the outputs of a map-called
pipeline, in some circumstances it is possible for the stage to attempt
to resolve its inputs before that pipeline has finished running.  In
particular, if the stage depends on an output of the pipeline that is
constant, it may start before that pipeline has even started.  However,
it still can't start until whatever determined the number of forks of
that pipeline is ready, and it must be able to resolve those forks.

Also fix a bunch of linter errors and improve some error messages.