Fix more nested map call cases.
In this case, the issue is when the nested map calls are both mapping
over the output of a non-mapped stage.

This also adds a new unit test case to run a pipeline.  It's implemented
as a unit test rather than an integration test so as to enable code
coverage instrumentation.