Do not treat an empty map call as single disabled.
It's not quite the same.  We want the result to be [] rather than
[null].

Also correct computation of forking dimension for pipeline dynamic merges
where the merged value doesn't depend on the inputs.

Also fix a few cases of fork ID aliasing, and other typos which caused
issues in some cases.