Updates to vscode extension.
Clean up execution of mro tool using promisify to simplify error
handling.

Use void instead of await for error reporting so that the function
doesn't report that it's still running until the user dismisses the
error message.

Write to the process using Readable.pipe, so that writing happens
asynchronously.

Move launch config to workspace root, so that it actually works.  Kind
of, anyway.

Update lint config, and typescript config to es2020.