Store timestamps in RFC 3339 form, with time zone.
Specifically applicable to the start/end times for stage jobinfos and
VDR reports.  Otherwise, particularly in cluster mode, we can run into
problems if the stage is seeing a different time zone from what the mrp
process sees, resulting in nonsense VDR results such as stages appearing
to complete before they started.  This also now preserves sub-second
resolution on these timestamps.

The tricky bit is that we want to still be able to parse metadata from
old pipestances, so we need to fall back on the old format for parsing.