Remove gulp, switch from npm to yarn.
Gulp seems to not be getting much maintenance, resulting in spurious
GitHub security warnings for transitive dependencies.  Switch to just
using a Makefile, which also aligns the non-bazel build better with the
bazel build, which doesn't use gulp either.

Switch from npm to yarn because the lockfile format is considerably more
stable and compact, and it's easier to review for duplication.