Generate CallName() and MroFileName() methods.
On Args objects generated for callables.  This is useful for programs
generating .mro files to serve as entry points for running a pipeline,
which will generally need look something like

    @include "pipeline.mro"

    call PIPELINE(
        ...,
    )

These methods allow such a call to be generated by
ast_builder.MakeCallAst, without any additional arguments.

Add bazel rules for generating code with mro2go.  This is a little more
complex than is feasible with a genrule because we need to pick up
transitive mro sources and set MROPATH properly.

Update the generated code comment to use the canonical path to the mro
file, to ensure consistency between different ways of running the tool.