Warn on bad mro2go-generated filenames
`mro2go` gets the file name of the generated go file by swapping the
extension of the MRO file, e.g. `foo.mro` → `foo.go`.  This can be
a problem for MRO files using the `_` prefix convention, because
`_foo.mro` will result in `_foo.go`, and lots of Go tooling (e.g. `go
build`) ignores files that start with an underscore.

We might consider trying to rename the files to make them work anyway,
but in the mean time we can at least log a warning to make it easier for
the user to figure out what's going wrong.