feat: report startup failures to enterprise service
If `mrp` is configured to register itself with an enterprise service, it
will report failures and error messages to it.  However, it would not
report failures during startup, e.g. failure to parse the invocation
source file or to create the pipestance directory, with the result that
enterprise would either not know anything at all about the mrp process
or would report that it terminated but not be able to provide any
further information to the user.

This change fixes that gap in error reporting.  It requires populating
some fields of the pipestanceBox state holder a bit earlier in the
startup process, and then attempting to report errors to enterprise (to
the extent possible) if there is a failure before the run loop starts.

It also requires passing back errors from instantiating the runtime,
so they can be reported, rather than exiting immediately.