feat: make SerializeState cancellable
SerializeState is called during post-processing, but also from the web
UI.  Under some circumstances, it can take a very long time to serialize
the data, so it's entirely possible for requests to time out before
they're returned.  Worse, clients may retry, which can result in a
pile-up.  When a request times out, we want to abort collection of that
data.

While we're at it, add more tasks and regions for tracing the process.

Note this includes breaking changes to the public go API.