fix: panic in SerializePerf if context is cancelled
Due to the changes in 541bc9f4593dc55fb51a4cc9592058a6a6f075eb,
it is now possible for some of the serialize methods to return nil if
the context is cancelled. This can result in a nil reference panic in a
few cases if the caller isn't prepared for that case.