mrjob: restore vmem limit after launching job
When --monitor is enabled, we've been seeing issues with errors like
runtime/cgo: out of memory in thread_start

This appears to be due to setting the vmem limit too low in some cases.
Once we've launched the subprocess, restore the vmem limit to what it
was originally, to avoid this sort of problem; monitor is supposed to be
for the stage code, not mrjob itself.