test: fix TestUnquoteFormat for go 1.22
Go 1.22 changed how the json encoder handles \b and \f, so one of the
test cases was failing.  The MRO encoder has been updated to match the
behavior of go 1.22, and that test case has been disabled when running
under go 1.21.

Update the CI to lock the go version, so we don't get random breakages
when new go versions are released, and to ensure we're testing both
supported versions of go.