feat: Add a martian.clear() method to set all outs to None
This change adds a new convince method to allow developers to set all
outs to None.

N.B. This wasn't added directly on the Record class
as that would have potentially conflicted with users setting an out with
the name `clear`. So instead we add the utility method to the general
martian module namespace.