Improve "not queued or running" error message.
Include the job ID for debugging purposes.

Use a somewhat different error message if the job had started up
successfully compared to if it had not.  The debugging steps are
generally different for jobs which started but then died unexpectedly
vs. ones that didn't start up at all, and the "not queued" part is
just confusing in the case of jobs which had started up.